Setup TDoc fetching and chunking in controller
We set up another DBClient and AutoFetcher instance in the controller.
Maintain two parallel instances for "reason" i.e TDoc and for "context" i.e specs.
TDocs when unzipped can often be in .doc form, so need to convert that.
Set up a helper function to do it using libreoffice (soffice command)