Implement TDoc DB and integrate with controller (#18)
* Update DBClient to handle multiple collections.

* Update autofetcher to have default behavior when no params specified

* Set up skeleton for frontend action to update tdoc collection n db

* Adjust heading/table detection for MAC

If text in the part does not have a style, the old way would fail.
So we check the class instance more explicitly

* Setup TDoc fetching and chunking in controller

We set up another DBClient and AutoFetcher instance in the controller.
Maintain two parallel instances for "reason" i.e TDoc and for "context" i.e specs.
TDocs when unzipped can often be in .doc form, so need to convert that.
Set up a helper function to do it using libreoffice (soffice command)

* Add extra document properties metadata into MAC