get precision and recall per file based on misalignment
Note a couple of things:
for each chunk (which Im defining as a section wide chunk), I get all the references in it\
 using the ReferenceExtractor.
 I'm calling this set the true_refs.

Then, I run the semantic search/ basic retriever using the chunk itself \
(again defined as a section wide chunk, the same chunk as in step 1) as the query.

Then, I use the ReferenceExtractor to pull out all the refs in the documents from the returned docs.
We'll called this set retrieved_refs
I also have a list of all the section names in the document itself.
Let's call this set all_sections_in_documents.